+++ pkg/pallets/constraints.go
@@ -2,27 +2,25 @@ package pallets
 
 // AttachedResource
 
-func NewAttachedResource[Resource any](
-	resource Resource, source []string,
-) AttachedResource[Resource] {
-	return AttachedResource[Resource]{
-		Resource: resource,
-		Source:   source,
-	}
-}
-
+// attachResources attaches the specified source to all of the specified resources.
 func attachResources[Resource any](
 	resources []Resource, source []string,
 ) (attached []AttachedResource[Resource]) {
 	attached = make([]AttachedResource[Resource], 0, len(resources))
 	for _, resource := range resources {
-		attached = append(attached, NewAttachedResource(resource, source))
+		attached = append(attached, AttachedResource[Resource]{
+			Resource: resource,
+			Source:   source,
+		})
 	}
 	return attached
 }
 
 // ResourceConflict
 
+// CheckResourcesConflicts identifies all resource conflicts between the first list of resources and
+// the second list of resources. It does not identify resource conflicts within the first list of
+// resources, nor within the second list of resources.
 func CheckResourcesConflicts[Resource ConflictChecker[Resource]](
 	first []AttachedResource[Resource], second []AttachedResource[Resource],
 ) (conflicts []ResourceConflict[Resource]) {
@@ -42,6 +40,8 @@ func CheckResourcesConflicts[Resource ConflictChecker[Resource]](
 
 // MissingResourceDependency
 
+// CheckResourcesDependencies identifies all unsatisfied resource dependencies between the provided
+// list of resource requirements and the provided list of resources.
 func CheckResourcesDependencies[Resource DependencyChecker[Resource]](
 	required []AttachedResource[Resource], provided []AttachedResource[Resource],
 ) (missingDeps []MissingResourceDependency[Resource]) {